backend/requirements.txt
# ========================================
# AI Ops Engineer - Backend Dependencies
# ========================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# AI & LLM
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
google-generativeai==0.3.2

# Vector Database & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2

# Utilities
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# PDF Generation
reportlab==4.0.8
weasyprint==60.1

# Email (optional)
aiosmtplib==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
